modularised TyreSetInfo and TyreSetHistoryEntry modularised WarningPenaltyEntry added unit tests
