+++ app.py
@@ -24,7 +24,6 @@
 
 import argparse
 import socket
-import sys
 import threading
 import time
 import webbrowser
